testing

+++ jsonP_parser.h
@@ -72,18 +72,18 @@ class jsonP_parser
 protected:
 	std::string json_str;
 	char * json = nullptr;
-	unsigned long json_length;
-	unsigned long index;
+	unsigned int json_length;
+	unsigned int index;
 	bool look_for_key;
-	unsigned long value_start;
+	unsigned int value_start;
 	
 	// STUFF for new modle to rid std::map and std::new for each element
 	byte * stack_buf;
-	unsigned long stack_buf_sz;
-	unsigned long stack_i;
+	unsigned int stack_buf_sz;
+	unsigned int stack_i;
 	byte * data;
-	unsigned long data_sz;
-	unsigned long data_i;
+	unsigned int data_sz;
+	unsigned int data_i;
 
 	//parse options
 	unsigned short options;
@@ -99,7 +99,7 @@ protected:
 
 	//error stuff
 	std::string error_string;
-	long error_index;
+	int error_index;
 	std::string error_json_snip;
 	
 	//virtual parse commands
@@ -114,30 +114,30 @@ protected:
 
 	virtual void parse_key();
 	virtual element_type parse_numeric();
-	virtual unsigned long parse_array();
-	virtual unsigned long parse_object();
+	virtual unsigned int parse_array();
+	virtual unsigned int parse_object();
 	virtual void parse_value();
 	
 	void set_error(std::string error);
 
-void test_parse_object(unsigned long);
-void test_parse_array(unsigned long);
+void test_parse_object(unsigned int);
+void test_parse_array(unsigned int);
 
 
 
 public:
 	jsonP_parser() = default;
 	jsonP_parser(std::string & json, unsigned short options = 0);
-	jsonP_parser(char * json, unsigned long, unsigned short options = 0);
+	jsonP_parser(char * json, unsigned int, unsigned short options = 0);
 	
 	~jsonP_parser();
 
 	jsonP_json * parse();
 	jsonP_json * parse(std::string & json);
-	jsonP_json * parse(char * json, unsigned long);
+	jsonP_json * parse(char * json, unsigned int);
 
 	std::string get_error_string() { return error_string; }
-	long get_error_index() { return error_index; }
+	int get_error_index() { return error_index; }
 	std::string get_error_snip(int chars_before, int chars_after);
 	
 	parse_stats get_parse_stats() { return stats; }
